Moved loadout view to loadout feature
