+++ src/app/features/Events/Event/EventLoadout.js
@@ -3,9 +3,8 @@ import PropTypes from 'prop-types'
 
 import Button from '@material-ui/core/Button'
 
-import LoadoutView from 'app/shared/components/Views/Loadout/LoadoutView'
+import { LoadoutView } from 'app/features/Loadouts'
 import ConfirmDeleteDialog from 'app/shared/components/Cards/ConfirmDeleteDialog'
-import LoadoutContextProvider from 'app/features/Loadouts/Loadout/LoadoutContextProvider'
 
 const EventLoadout = ({ event, activeUserIndex, removeLoadout }) => {
 	let [ activeDialog, setActiveDialog ] = useState() 
@@ -34,9 +33,7 @@ const EventLoadout = ({ event, activeUserIndex, removeLoadout }) => {
 				<div style={ {marginTop: '24px'} }/>
 			}
 					
-			<LoadoutContextProvider loadout={ activeLoadout } >
-				<LoadoutView />
-			</LoadoutContextProvider>
+			<LoadoutView loadout={ activeLoadout }/>
 
 			<ConfirmDeleteDialog 
 				verb='Remove'
