+++ src/app/features/Loadouts/Loadout/GearList/LoadoutGearList.js
@@ -6,7 +6,7 @@ import AddButton from 'app/shared/components/Buttons/AddButton'
 import LoadoutGear from './Gear/LoadoutGear'
 import AddArmoryItemDialog from '../AddArmoryItemDialog/AddArmoryItemDialog'
 
-import database from '../../../../../../firebase/database'
+import database from '../../../../../firebase/database'
 
 import './LoadoutGearList.css'
 import LoadoutContext from 'app/features/Loadouts/Loadout/LoadoutContext'
